Move append_vec::test_utils into its own file
